fixed things. To do still: more oracles, more matroid types, customisable algorithms.

+++ src/main/Java/Randomiser/RandomiserFactory.java
@@ -6,5 +6,5 @@ package Java.Randomiser;
 public interface RandomiserFactory {
     Randomiser createRandomiser();
 
-    void update(String randomiserType, long[] buildData);
+    void update(String randomiserType, double[] buildData);
 }
